src/test/java/kata/supermarket/WeighedProductTest.java
package kata.supermarket;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.math.BigDecimal;
import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

class WeighedProductTest {

    @ParameterizedTest
    @MethodSource
    void itemFromWeighedProductHasExpectedUnitPrice(String pricePerKilo, String weightInKilos, String expectedPrice, String code) {
        final WeighedProduct weighedProduct = new WeighedProduct(new BigDecimal(pricePerKilo), code);
        final Item weighedItem = weighedProduct.weighing(new BigDecimal(weightInKilos));
        assertEquals(new BigDecimal(expectedPrice), weighedItem.price());
    }

    static Stream<Arguments> itemFromWeighedProductHasExpectedUnitPrice() {
        return Stream.of(
                Arguments.of("100.00", "1.00", "100.00", "12345"),
                Arguments.of("100.00", "0.33333", "33.33", "23456"),
                Arguments.of("100.00", "0.33335", "33.34", "34567"),
                Arguments.of("100.00", "0", "0.00", "45678")
        );
    }

}